scrape_examples

Function test_many

pub fn test_many()

Examples found in repository?

examples/check-many-1.rs (line 3)

1fn main() {
2    // all examples have same line count
3    scrape_examples::test_many();
4    scrape_examples::test_many();
5    scrape_examples::test_many();
6    scrape_examples::test_many();
7    scrape_examples::test_many();
8    scrape_examples::test_many();
9    scrape_examples::test_many();
10    scrape_examples::test_many();
11    scrape_examples::test_many();
12    scrape_examples::test_many();
13}

Hide additional examples

examples/check-many-4.rs (line 3)

1fn main() {
2    // all examples have same line count
3    scrape_examples::test_many();
4    scrape_examples::test_many();
5    scrape_examples::test_many();
6    scrape_examples::test_many();
7    scrape_examples::test_many();
8    scrape_examples::test_many();
9    scrape_examples::test_many();
10    scrape_examples::test_many();
11    scrape_examples::test_many();
12    scrape_examples::test_many();
13}

examples/check-many-5.rs (line 3)

1fn main() {
2    // all examples have same line count
3    scrape_examples::test_many();
4    scrape_examples::test_many();
5    scrape_examples::test_many();
6    scrape_examples::test_many();
7    scrape_examples::test_many();
8    scrape_examples::test_many();
9    scrape_examples::test_many();
10    scrape_examples::test_many();
11    scrape_examples::test_many();
12    scrape_examples::test_many();
13}

examples/check-many-6.rs (line 3)

1fn main() {
2    // all examples have same line count
3    scrape_examples::test_many();
4    scrape_examples::test_many();
5    scrape_examples::test_many();
6    scrape_examples::test_many();
7    scrape_examples::test_many();
8    scrape_examples::test_many();
9    scrape_examples::test_many();
10    scrape_examples::test_many();
11    scrape_examples::test_many();
12    scrape_examples::test_many();
13}

examples/check-many-7.rs (line 3)

1fn main() {
2    // all examples have same line count
3    scrape_examples::test_many();
4    scrape_examples::test_many();
5    scrape_examples::test_many();
6    scrape_examples::test_many();
7    scrape_examples::test_many();
8    scrape_examples::test_many();
9    scrape_examples::test_many();
10    scrape_examples::test_many();
11    scrape_examples::test_many();
12    scrape_examples::test_many();
13}

examples/check-many-2.rs (line 3)

2fn main() {
3    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
4    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
5    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
6    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
7    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
8    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
9    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
10    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
11    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
12    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
13}

Additional examples can be found in:

examples/check-many-3.rs