Function scrape_examples::test_many

pub fn test_many()

Examples found in repository?

examples/check-many-1.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // all examples have same line count
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
}

Hide additional examples

examples/check-many-4.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // all examples have same line count
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
}

examples/check-many-5.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // all examples have same line count
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
}

examples/check-many-6.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // all examples have same line count
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
}

examples/check-many-7.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // all examples have same line count
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
    scrape_examples::test_many();
}

examples/check-many-2.rs (line 3)

1
2
3
4
5
6
7
8
9
10
11
12
13

fn main() {
    // ignore-tidy-linelength
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
    scrape_examples::test_many(); /* Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. */
}

Additional examples can be found in:

examples/check-many-3.rs